initial python metadata bindings work
Starting simple, `scene_lux` is the only thing bound right now.
More `Argus` types need to be bound before the other things work.

Trickiest part was the `from_buffer` casting as how this is done
(between GObject bindings and Pybind) is completely undocumented.

Nvidia used `reinterpret_cast` of the python object's hash which is
apparently the address of the actual GObject. We take the same
approach but do  not require anybody to call `hash` from python
because pybind has `py::hash`